[Fix] Reduce raycasting area to improve performance
Check the measurements that are above the ray.
If the measurement exists above the ray, the ray is not casted. (break the for loop)
This is to prevent unnecessary height correction to the static obstacles.
The main cause of issue was the partial observability (we cannot measure inside the obstacle).